Created Navbar component
